restructured directory structure
